[REFACTOR] 즐겨찾기 등록, 조회, 삭제의 validation 추가 (김세현)

+++ src/main/java/com/ae/chaebbiSpring/api/BookmarkApiController.java
@@ -1,5 +1,6 @@
 package com.ae.chaebbiSpring.api;
 
+import com.ae.chaebbiSpring.config.BaseResponse;
 import com.ae.chaebbiSpring.domain.Bistro;
 import com.ae.chaebbiSpring.domain.Bookmark;
 import com.ae.chaebbiSpring.domain.User;
@@ -20,6 +21,9 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.ae.chaebbiSpring.config.BaseResponseStatus.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -30,37 +34,69 @@ public class BookmarkApiController {
 
     //7-1
     @PostMapping("api/bookmark")
-    public CreateBookmarkResponseDto createBookmarkResponse(@AuthenticationPrincipal String userId,
-                                                         @RequestBody @Valid BookmarkRequestDto request) {
+    public BaseResponse<CreateBookmarkResponseDto> createBookmarkResponse(@AuthenticationPrincipal String userId,
+                                                                          @RequestBody @Valid BookmarkRequestDto request) {
+        if(userId == null) {
+            return new BaseResponse<>(EMPTY_JWT);
+        }
         User user = userService.findOne(Long.valueOf(userId));
+        if (request.getBistroId() == null || request.getBistroId().equals("")){
+            return new BaseResponse<>(POST_BOOKMARK_NO_BISTRO_ID);
+        }
+        List<Bistro> restaurant = bookmarkService.findBookmark(Long.valueOf(userId));
+        List<Long> restaurantId = restaurant.stream().map(Bistro::getId).collect(Collectors.toList());
+        Long count = restaurantId.stream().filter(m-> request.getBistroId().equals(m)).count();
+        if(count > 0 ){
+            return new BaseResponse<>(POST_BOOKMARK_PRESENT_BISTRO);
+        }
         Bistro bistro = bistroService.findOne(request.getBistroId());
         Bookmark bookmark = Bookmark.createBookmark(user, bistro);
         Long id = bookmarkService.create(bookmark);
 
-        return new CreateBookmarkResponseDto(id.intValue());
+        return new BaseResponse<>(new CreateBookmarkResponseDto(id.intValue()));
     }
 
     //7-2
     @GetMapping("api/bookmarklist")
-    public ResResponse bookmarkList(@AuthenticationPrincipal String userId) {
+    public BaseResponse<ResResponse> bookmarkList(@AuthenticationPrincipal String userId) {
+        if(userId == null) {
+            return new BaseResponse<>(EMPTY_JWT);
+        }
         List<Bistro> restaurant = bookmarkService.findBookmark(Long.valueOf(userId));
         List<RestaurantResponseDto> restaurantDtos = new ArrayList<>();
 
         for(Bistro bistro: restaurant) {
-            restaurantDtos.add(new RestaurantResponseDto(bistro.getCategory(), bistro.getName(),
+            restaurantDtos.add(new RestaurantResponseDto(bistro.getId(), bistro.getCategory(), bistro.getName(),
                     bistro.getRAddr(), bistro.getLAddr(),
                     bistro.getTel(), bistro.getLa(), bistro.getLo()));
         }
-        return new ResResponse(restaurantDtos.size(), restaurantDtos);
+        if(restaurant.size() > 0){
+            return new BaseResponse<>(new ResResponse(restaurantDtos.size(), restaurantDtos));
+
+        }else return new BaseResponse<>(POST_BOOKMARK_LIST_EMPTY);
 
     }
 
     //7-3
     @DeleteMapping("api/del/bookmark")
-    public CreateBookmarkResponseDto deleteBookmark(@AuthenticationPrincipal String userId,
+    public BaseResponse<CreateBookmarkResponseDto> deleteBookmark(@AuthenticationPrincipal String userId,
                               @RequestBody @Valid BookmarkRequestDto request){
-        Long bistroId = bookmarkService.deleteBookmark(Long.valueOf(userId), request.getBistroId());
-        return new CreateBookmarkResponseDto(bistroId.intValue());
+        if(userId == null) {
+            return new BaseResponse<>(EMPTY_JWT);
+        }
+
+        List<Bistro> restaurant = bookmarkService.findBookmark(Long.valueOf(userId));
+        Long count =restaurant.stream().map(Bistro::getId).collect(Collectors.toList())
+                .stream().filter(
+                m-> request.getBistroId().equals(m))
+                .count();
+        if(count > 0 ){
+            Long bistroId = bookmarkService.deleteBookmark(Long.valueOf(userId), request.getBistroId());
+            return new BaseResponse<>(new CreateBookmarkResponseDto(bistroId.intValue()));
+
+        } else return new BaseResponse<>(POST_BOOMARK_THERE_NO_BISTRO);
+
+
 
     }
